Incorporate IE get xDomain fix.
The commit integrates the changes in [1] which are in turn based on this
video.js pull request [2]. This should fix a problem whereby cross-domain
Ajax GET requests from within video.js did not work for Internet
Explorer < v10.

This should mean that subtitles (which are hosted on a sub-domain of
`futurelearn.com` in production, i.e. `ugc.futurelearn.com`) work with
the video.js player on earlier versions of Internet Explorer.

[1] https://github.com/Futurelearn/video.js/compare/v4.5.2-with-IE-get-xDomain-fix
[2] https://github.com/videojs/video.js/pull/1095