Introduce local SWF back in.
We don’t want to depend on a CDN that may be down or be blocked
in other countries.

The more CDNs and services we depend on the more likely that one
of them may have problems. Include the SWF back in the gem so we
have the option to use it.