Allow videojs to parse vtt with headers
Previously VideoJS would not take into account whether a vtt file had
headers or not. Instead making the assumption that except for the first
line, the rest of the content for the file was part of a valid subtitle
file.

This commit allows it to parse the headers successfully. This is only a
temporary fix and once all subtitles on the platform are using the vtt
format, we will update our version of videojs and will therefore no
longer need this hack.